--- a/lttoolbox/reusable_state.cc
+++ b/lttoolbox/reusable_state.cc
@@ -1,5 +1,6 @@
 #include <reusable_state.h>
 #include <climits>
+#include <algorithm>
 
 #define WalkBack(var, pos, block) { \
   size_t index = pos;               \