This PR implements and alternative to `State` that builds a graph of output paths rather than copying a collection of vectors. The result is that `step()` is faster because it's not doing any allocations and `filterFinals()` is slower because it has to collect the path rather than just iterating over an array. Initial experimentation suggests this is an overall speedup, but the constants in `resuable_state.h` should probably be tuned empirically.

I currently have it as a separate class rather than directly modifying `State` in case there ends up being any ABI breakage or if this turns out to be better for some processes but not others.