@mr-martian 2024-07-26
Having now implemented this for analysis, I get the following results:

| metric | old -oci | new -oci | old -eng | new -eng |
|---|---|---|---|---|
| input words | 95K | 95K | 68K | 68K |
| memory | 440 MB | 439 MB | 34 MB | 33 MB |
| time | 2.1 s | 1.7 s | 0.6 s | 0.4 s |
| speedup | | 19% | | 33% |