@mr-martian 2024-07-26
Restoring the old ordering is just
```diff
diff --git a/lttoolbox/reusable_state.cc b/lttoolbox/reusable_state.cc
index aa55c95..1126d71 100644
--- a/lttoolbox/reusable_state.cc
+++ b/lttoolbox/reusable_state.cc
@@ -241,11 +241,17 @@ void ReusableState::extract(size_t pos, UString& result, double& weight,
   }
 }
 
+bool comp_pair(const std::pair<double, UString>& a,
+               const std::pair<double, UString>& b)
+{
+  return a.first < b.first;
+}
+
 void NFinals(std::vector<std::pair<double, UString>>& results,
              size_t maxAnalyses, size_t maxWeightClasses)
 {
   if (results.empty()) return;
-  sort(results.begin(), results.end());
+  sort(results.begin(), results.end(), comp_pair);
   if (maxAnalyses < results.size()) {
     results.erase(results.begin()+maxAnalyses, results.end());
   }
```
which has a negligible impact on the runtime. I leave the question open to bikeshedding.