@mr-martian 2024-07-30
Having tried block sizes of `1<<N` for `N=1..8,16`, my tentative conclusion is that changing it doesn't meaningfully alter the runtime numbers listed above.