@mr-martian 2024-09-05
turns out I actually just forgot to check the dirty bit