@mr-martian 2024-09-11
Unfortunately, the memory usage on this explodes for nno-nob, probably due to regexes (see #167).